Disable Chromatic snapshots for the whole react-openlayers-fiber package
