+++ typescript/react-openlayers-fiber/src/__stories__/basic.tsx
@@ -2,10 +2,6 @@ import React from "react";
 import { useInterval } from "react-use";
 import { Map } from "../map";
 
-export default {
-  title: "react-openlayers-fiber/Other Examples",
-};
-
 /**
  * In this story we prefix the props with initial for the view, so it should be initially at these values
  * and allow free movement after, never reseting the view position
@@ -66,4 +62,3 @@ export const BasicBoth = () => {
     </Map>
   );
 };
-BasicBoth.parameters = { chromatic: { disableSnapshot: true } };
