+++ typescript/react-openlayers-fiber/src/__stories__/cluster.tsx
@@ -73,9 +73,3 @@ export const Cluster = () => {
     </div>
   );
 };
-
-Cluster.parameters = {
-  // disables Chromatic's snapshotting on a story level
-  // See https://www.chromatic.com/docs/ignoring-elements
-  chromatic: { disableSnapshot: true },
-};
