+++ typescript/react-openlayers-fiber/src/__stories__/draw-and-modify-features.tsx
@@ -72,7 +72,3 @@ export const DrawAndModifyFeatures = () => {
     </>
   );
 };
-
-DrawAndModifyFeatures.parameters = {
-  chromatic: { disableSnapshot: true },
-};
