+++ typescript/react-openlayers-fiber/src/__stories__/draw-shapes.tsx
@@ -96,7 +96,3 @@ export const DrawShapes = () => {
     </>
   );
 };
-
-DrawShapes.parameters = {
-  chromatic: { disableSnapshot: true },
-};
