+++ typescript/react-openlayers-fiber/src/__stories__/dynamic-data.tsx
@@ -73,9 +73,3 @@ export const DynamicData = () => {
     </Map>
   );
 };
-
-DynamicData.parameters = {
-  // disables Chromatic's snapshotting on a story level
-  // See https://www.chromatic.com/docs/ignoring-elements
-  chromatic: { disableSnapshot: true },
-};
