+++ typescript/react-openlayers-fiber/src/__stories__/icon.tsx
