+++ typescript/react-openlayers-fiber/src/__stories__/image-filter.tsx
