+++ typescript/react-openlayers-fiber/src/__stories__/kitchen-sink.tsx
