+++ typescript/react-openlayers-fiber/src/__stories__/mapbox-vector-tiles.tsx
