+++ typescript/react-openlayers-fiber/src/__stories__/ol-examples.stories.tsx
@@ -2,29 +2,30 @@ import { Map } from "../map";
 
 import "ol/ol.css";
 
-export * from "./accessible.stories";
-export * from "./bing.stories";
-export * from "./cluster.stories";
-export * from "./draw-and-modify-features.stories";
-export * from "./draw-shapes.stories";
-export * from "./dynamic-data.stories";
-export * from "./geojson.stories";
-export * from "./icon.stories";
-export * from "./image-filter.stories";
-export * from "./mapbox-vector-tiles.stories";
-export * from "./popup.stories";
-export * from "./reprojection-image.stories";
-export * from "./reprojection-wgs84.stories";
-export * from "./select-features.stories";
-export * from "./static-image.stories";
-export * from "./turf.stories";
-export * from "./vector-tiles.stories";
-export * from "./wms-tiled.stories";
-export * from "./xyz-retina.stories";
-export * from "./xyz.stories";
-export * from "./zoomify.stories";
+export * from "./accessible";
+export * from "./bing";
+export * from "./cluster";
+export * from "./draw-and-modify-features";
+export * from "./draw-shapes";
+export * from "./dynamic-data";
+export * from "./geojson";
+export * from "./icon";
+export * from "./image-filter";
+export * from "./mapbox-vector-tiles";
+export * from "./popup";
+export * from "./reprojection-image";
+export * from "./reprojection-wgs84";
+export * from "./select-features";
+export * from "./static-image";
+export * from "./turf";
+export * from "./vector-tiles";
+export * from "./wms-tiled";
+export * from "./xyz-retina";
+export * from "./xyz";
+export * from "./zoomify";
 
 export default {
   title: "react-openlayers-fiber/OL Examples",
   component: Map,
+  parameters: { chromatic: { disableSnapshot: true } },
 };
