+++ typescript/react-openlayers-fiber/src/__stories__/other-examples.stories.tsx
@@ -2,14 +2,15 @@ import { Map } from "../map";
 
 import "ol/ol.css";
 
-export * from "./basic.stories";
-export * from "./dynamic-style.stories";
-export * from "./dynamic.stories";
-export * from "./kitchen-sink.stories";
-export * from "./performance.stories";
-export * from "./retina.stories";
+export * from "./basic";
+export * from "./dynamic-style";
+export * from "./dynamic";
+export * from "./kitchen-sink";
+export * from "./performance";
+export * from "./retina";
 
 export default {
   title: "react-openlayers-fiber/Other Examples",
   component: Map,
+  parameters: { chromatic: { disableSnapshot: true } },
 };
