+++ typescript/react-openlayers-fiber/src/__stories__/performance.tsx
@@ -55,9 +55,3 @@ export const Performance = () => {
     </Map>
   );
 };
-
-Performance.parameters = {
-  // disables Chromatic's snapshotting on a story level
-  // See https://www.chromatic.com/docs/ignoring-elements
-  chromatic: { disableSnapshot: true },
-};
