+++ typescript/react-openlayers-fiber/src/__stories__/popup.tsx
@@ -82,7 +82,3 @@ export const Popup = () => {
     </div>
   );
 };
-
-Popup.parameters = {
-  chromatic: { disableSnapshot: true },
-};
