+++ typescript/react-openlayers-fiber/src/__stories__/reprojection-image.tsx
@@ -59,9 +59,3 @@ export const ReprojectionImage = () => {
     </>
   );
 };
-
-ReprojectionImage.parameters = {
-  // disables Chromatic's snapshotting on a story level
-  // See https://www.chromatic.com/docs/ignoring-elements
-  chromatic: { disableSnapshot: true },
-};
