+++ typescript/react-openlayers-fiber/src/__stories__/retina.tsx
@@ -16,7 +16,3 @@ export const Retina = () => (
     </olLayerTile>
   </Map>
 );
-
-Retina.parameters = {
-  chromatic: { disableSnapshot: true },
-};
