+++ typescript/react-openlayers-fiber/src/__stories__/static-image.tsx
