+++ typescript/react-openlayers-fiber/src/__stories__/turf.tsx
@@ -67,7 +67,3 @@ export const Turf = () => {
     </Map>
   );
 };
-
-Turf.parameters = {
-  chromatic: { disableSnapshot: true },
-};
