+++ typescript/react-openlayers-fiber/src/__stories__/vector-tiles.tsx
