+++ typescript/react-openlayers-fiber/src/__stories__/wms-tiled.tsx
@@ -20,7 +20,3 @@ export const TiledWMS = () => {
     </Map>
   );
 };
-
-TiledWMS.parameters = {
-  chromatic: { disableSnapshot: true },
-};
