+++ typescript/react-openlayers-fiber/src/__stories__/xyz-retina.tsx
@@ -27,5 +27,3 @@ export const XYZRetina = () => {
     </Map>
   );
 };
-
-XYZRetina.parameters = { chromatic: { disableSnapshot: true } };
