+++ typescript/react-openlayers-fiber/src/__stories__/zoomify.tsx
